fix(update): Lazy-link SHA256 so unused Update sketches keep FLASH
Move SHA256 into its own TU and bind it only from setSHA256(), matching
the SHA512 pattern so --gc-sections can drop mbedtls/PSA SHA256.
