+++ CMakeLists.txt
@@ -255,6 +255,7 @@ set(ARDUINO_LIBRARY_Ticker_SRCS libraries/Ticker/src/Ticker.cpp)
 
 set(ARDUINO_LIBRARY_Update_SRCS
   libraries/Update/src/Updater.cpp
+  libraries/Update/src/UpdaterSHA256.cpp
   libraries/Update/src/UpdaterSHA512.cpp
   libraries/Update/src/HttpsOTAUpdate.cpp
   libraries/Update/src/Updater_Signing.cpp)
