+++ libraries/Update/src/Updater.cpp
@@ -10,94 +10,18 @@
 #include "esp_ota_ops.h"
 #include "esp_image_format.h"
 #include "HEXBuilder.h"
+#ifndef UPDATE_NOCRYPT
 #include "mbedtls/build_info.h"
 #if MBEDTLS_VERSION_MAJOR >= 4
 #include "psa/crypto.h"
 #else
-#include "mbedtls/sha256.h"
-#ifndef UPDATE_NOCRYPT
 #include "mbedtls/aes.h"
-#endif /* UPDATE_NOCRYPT */
 #endif /* MBEDTLS_VERSION_MAJOR >= 4 */
+#endif /* UPDATE_NOCRYPT */
 
-// Streaming SHA-256 context used only when setSHA256() is called.
-// Prefers IDF/mbedTLS backends over SHA2Builder to avoid linking the full
-// software SHA-224/256/384/512 implementation into every Update sketch.
-// Optional SHA-512 lives in UpdaterSHA512.cpp and is bound from setSHA512().
-struct UpdateSHA256Context {
-#if MBEDTLS_VERSION_MAJOR >= 4
-  psa_hash_operation_t op;
-#else
-  mbedtls_sha256_context ctx;
-#endif
-  uint8_t expected[32];
-};
-
-static bool update_sha256_begin(UpdateSHA256Context *ctx) {
-#if MBEDTLS_VERSION_MAJOR >= 4
-  psa_status_t psa_ret = psa_crypto_init();
-  if (psa_ret != PSA_SUCCESS) {
-    log_e("PSA crypto init failed: %d", (int)psa_ret);
-    return false;
-  }
-  ctx->op = psa_hash_operation_init();
-  psa_ret = psa_hash_setup(&ctx->op, PSA_ALG_SHA_256);
-  if (psa_ret != PSA_SUCCESS) {
-    log_e("PSA hash setup failed: %d", (int)psa_ret);
-    return false;
-  }
-  return true;
-#else
-  mbedtls_sha256_init(&ctx->ctx);
-  if (mbedtls_sha256_starts(&ctx->ctx, 0) != 0) {
-    mbedtls_sha256_free(&ctx->ctx);
-    return false;
-  }
-  return true;
-#endif
-}
-
-static bool update_sha256_update(UpdateSHA256Context *ctx, const uint8_t *data, size_t len) {
-  if (!ctx || !data || len == 0) {
-    return true;
-  }
-#if MBEDTLS_VERSION_MAJOR >= 4
-  return psa_hash_update(&ctx->op, data, len) == PSA_SUCCESS;
-#else
-  return mbedtls_sha256_update(&ctx->ctx, data, len) == 0;
-#endif
-}
-
-static bool update_sha256_finish(UpdateSHA256Context *ctx, uint8_t out[32]) {
-  if (!ctx || !out) {
-    return false;
-  }
-#if MBEDTLS_VERSION_MAJOR >= 4
-  size_t hash_len = 0;
-  psa_status_t psa_ret = psa_hash_finish(&ctx->op, out, 32, &hash_len);
-  if (psa_ret != PSA_SUCCESS || hash_len != 32) {
-    psa_hash_abort(&ctx->op);
-  }
-  ctx->op = psa_hash_operation_init();
-  return psa_ret == PSA_SUCCESS && hash_len == 32;
-#else
-  int ret = mbedtls_sha256_finish(&ctx->ctx, out);
-  mbedtls_sha256_free(&ctx->ctx);
-  return ret == 0;
-#endif
-}
-
-static void update_sha256_abort(UpdateSHA256Context *ctx) {
-  if (!ctx) {
-    return;
-  }
-#if MBEDTLS_VERSION_MAJOR >= 4
-  psa_hash_abort(&ctx->op);
-  ctx->op = psa_hash_operation_init();
-#else
-  mbedtls_sha256_free(&ctx->ctx);
-#endif
-}
+// Optional SHA-256/SHA-512 live in UpdaterSHA256.cpp / UpdaterSHA512.cpp
+// and are bound from setSHA256() / setSHA512() so --gc-sections can drop
+// mbedtls/PSA hash implementations from Update sketches that never enable them.
 
 static const char *_err2str(uint8_t _error) {
   if (_error == UPDATE_ERROR_OK) {
@@ -170,7 +94,7 @@ UpdateClass::UpdateClass()
     _cryptKey(0), _cryptBuffer(0),
 #endif /* UPDATE_NOCRYPT */
     _buffer(0), _skipBuffer(0), _bufferLen(0), _size(0), _progress_callback(NULL), _progress(0), _command(U_FLASH), _partition(NULL), _sha256_ctx(NULL),
-    _sha256_valid(false), _sha512_ctx(NULL), _sha512_valid(false), _sha512Ops(NULL)
+    _sha256_valid(false), _sha512_ctx(NULL), _sha512_valid(false), _sha256Ops(NULL), _sha512Ops(NULL)
 #ifndef UPDATE_NOCRYPT
     ,
     _cryptMode(U_AES_DECRYPT_AUTO), _cryptAddress(0), _cryptCfg(0xf)
@@ -208,7 +132,9 @@ void UpdateClass::_reset() {
   }
 #endif /* UPDATE_SIGN */
 
-  _sha256FreeContext();
+  if (_sha256Ops) {
+    _sha256Ops->freeContext(_sha256_ctx);
+  }
   if (_sha512Ops) {
     _sha512Ops->freeContext(_sha512_ctx);
   }
@@ -289,12 +215,8 @@ bool UpdateClass::begin(size_t size, int command, int ledPin, uint8_t ledOn, con
   _error = 0;
   _target_md5 = emptyString;
   _md5 = MD5Builder();
-  _sha256FreeContext();
   _sha256_valid = false;
   memset(_sha256_result, 0, sizeof(_sha256_result));
-  if (_sha512Ops) {
-    _sha512Ops->freeContext(_sha512_ctx);
-  }
   _sha512_valid = false;
   memset(_sha512_result, 0, sizeof(_sha512_result));
 #ifndef UPDATE_NOCRYPT
@@ -729,7 +651,7 @@ bool UpdateClass::_writeBuffer() {
     _md5.add(_buffer, _bufferLen);
   }
   if (!_target_sha256_decrypted) {
-    if (!_sha256Update(_buffer, _bufferLen)) {
+    if (_sha256Ops && !_sha256Ops->update(_sha256_ctx, _buffer, _bufferLen)) {
       _abort(UPDATE_ERROR_SHA256);
       return false;
     }
@@ -806,7 +728,7 @@ bool UpdateClass::_writeBuffer() {
   }
   if (_target_sha256_decrypted) {
 #endif /* UPDATE_NOCRYPT */
-    if (!_sha256Update(_buffer, _bufferLen)) {
+    if (_sha256Ops && !_sha256Ops->update(_sha256_ctx, _buffer, _bufferLen)) {
       _abort(UPDATE_ERROR_SHA256);
       return false;
     }
@@ -897,47 +819,6 @@ bool UpdateClass::setMD5(
   return true;
 }
 
-bool UpdateClass::setSHA256(
-  const char *expected_sha256
-#ifndef UPDATE_NOCRYPT
-  ,
-  bool calc_post_decryption
-#endif /* UPDATE_NOCRYPT */
-) {
-  if (!expected_sha256 || strlen(expected_sha256) != 64 || !HEXBuilder::isHexString(expected_sha256, 64)) {
-    return false;
-  }
-  // Digest must cover the entire payload; reject once writing has started.
-  if (_progress > 0 || _bufferLen > 0) {
-    log_e("setSHA256 must be called before writing data");
-    return false;
-  }
-  if (!isRunning()) {
-    log_e("setSHA256 requires an active update (call begin first)");
-    return false;
-  }
-
-  UpdateSHA256Context *ctx = new (std::nothrow) UpdateSHA256Context;
-  if (!ctx) {
-    log_e("Failed to allocate SHA-256 context");
-    return false;
-  }
-  if (!update_sha256_begin(ctx)) {
-    delete ctx;
-    return false;
-  }
-
-  HEXBuilder::hex2bytes(ctx->expected, sizeof(ctx->expected), expected_sha256);
-  _sha256FreeContext();
-  _sha256_valid = false;
-  memset(_sha256_result, 0, sizeof(_sha256_result));
-  _sha256_ctx = ctx;
-#ifndef UPDATE_NOCRYPT
-  _target_sha256_decrypted = calc_post_decryption;
-#endif /* UPDATE_NOCRYPT */
-  return true;
-}
-
 String UpdateClass::sha256String(void) {
   if (!_sha256_valid) {
     return String();
@@ -954,47 +835,6 @@ void UpdateClass::sha256(uint8_t *result) {
   } else {
     memset(result, 0, sizeof(_sha256_result));
   }
-}
-
-void UpdateClass::_sha256FreeContext() {
-  if (_sha256_ctx) {
-    update_sha256_abort(static_cast<UpdateSHA256Context *>(_sha256_ctx));
-    delete static_cast<UpdateSHA256Context *>(_sha256_ctx);
-    _sha256_ctx = nullptr;
-  }
-}
-
-bool UpdateClass::_sha256Update(const uint8_t *data, size_t len) {
-  if (!_sha256_ctx) {
-    return true;
-  }
-  if (!update_sha256_update(static_cast<UpdateSHA256Context *>(_sha256_ctx), data, len)) {
-    log_e("SHA-256 update failed");
-    _sha256FreeContext();
-    return false;
-  }
-  return true;
-}
-
-bool UpdateClass::_sha256Finish() {
-  if (!_sha256_ctx) {
-    return false;
-  }
-  UpdateSHA256Context *ctx = static_cast<UpdateSHA256Context *>(_sha256_ctx);
-  uint8_t digest[32];
-  bool ok = update_sha256_finish(ctx, digest);
-  if (ok) {
-    ok = memcmp(ctx->expected, digest, sizeof(digest)) == 0;
-  }
-  // Streaming state is consumed by finish; drop the heap object either way.
-  delete ctx;
-  _sha256_ctx = nullptr;
-  if (!ok) {
-    _sha256_valid = false;
-    return false;
-  }
-  memcpy(_sha256_result, digest, sizeof(_sha256_result));
-  return true;
 }
 
 String UpdateClass::sha512String(void) {
@@ -1043,7 +883,7 @@ bool UpdateClass::end(bool evenIfRemaining) {
 
   bool sha256_used = _sha256_ctx != nullptr;
   if (sha256_used) {
-    if (!_sha256Finish()) {
+    if (!_sha256Ops || !_sha256Ops->finish(_sha256_ctx, _sha256_result, _sha256_valid)) {
       _abort(UPDATE_ERROR_SHA256);
       return false;
     }
